requirements.txt
# requirements.txt
fastapi
uvicorn
pydantic
openai==0.28
python-dotenv
httpx 